Add tests for VertoConversionError
